Implementación de la gestión de entrada de stock con los recursos de Filament
Se introducen los modelos StockEntry y StockEntryItem, las migraciones y los observadores para gestionar las entradas de inventario. Se añaden recursos, páginas, esquemas y tablas de Filament para operaciones CRUD en las entradas de inventario. Se actualiza el modelo de producto para gestionar los aumentos de inventario y se registran nuevos observadores en AppServiceProvider. También se incluyen actualizaciones de los recursos frontend relacionados y actualizaciones de dependencias.

+++ database/migrations/2025_12_29_125834_create_products_table.php
@@ -19,9 +19,9 @@ return new class extends Migration
                 ->nullable()
                 ->constrained('categories')
                 ->onDelete('restrict');
-            $table->integer('units_available');
-            $table->float('unit_cost');
-            $table->float('sales_price');
+            $table->unsignedInteger('units_available');
+            $table->decimal('unit_cost');
+            $table->decimal('sales_price',15,2);
             $table->string('attachment')->nullable();
             $table->softDeletes();
             $table->timestamps();
